+++ packages/react-canvas-fiber/src/utils/index.ts
@@ -1,2 +1,20 @@
 export { normalizeInsets } from './style'
 export { resolvePath2D } from './path2d'
+export type { Matrix2D, Point2D } from './transform'
+export {
+	IDENTITY_MATRIX,
+	applyToPoint,
+	estimateUniformScale,
+	invertMatrix,
+	multiplyMatrix,
+	parseTransform,
+	resolveOpacity,
+	resolveOverflowHidden,
+	resolveTransformOrigin,
+	resolveZIndex,
+	rotationMatrix,
+	scaleMatrix,
+	skewXMatrix,
+	skewYMatrix,
+	translationMatrix,
+} from './transform'